More consistent use of uint64.
git-svn-id: svn+ssh://truffle/home/andrew/repo/memview@105 6d4121a9-3ab8-48c6-8db9-97dca464e206

+++ Loader.C
@@ -241,8 +241,8 @@ Loader::loadFromLackey(int max_read)
 	}
 
 	uint64		addr;
-	int		size;
-	int		type;
+	uint64		size;
+	uint64		type;
 
 	char		*saveptr = 0;
 	char		*type_str;
@@ -374,7 +374,7 @@ public:
 	int count = myBlock->myEntries;
 	for (int i = 0; i < count; i++)
 	{
-	    unsigned long long addr = myBlock->myAddr[i];
+	    uint64 addr = myBlock->myAddr[i];
 	    myState->updateAddress(
 		    addr & theAddrMask,
 		    addr >> theSizeShift,
